requirements.txt
py-cord==2.0.0
python-dotenv
pycryptodome==3.10.1
requests
Pillow
